Add two socket options.
This adds two socket options to the listener socket. The first is the
KEEPALIVE option. This will help detect sooner than before if the
client goes away, which had to wait until a report needed to be sent
before detecting a missing client.

The second option is to disbale the EPIPE signal. Now we just get an
error code instead of a signal.