Remove syslog.
When manually shorting out the GPIO pin, the LED would stay on for a
surprisingly long time. Apparently, parsing the format string and
sending the UDP message took milliseconds, which was even visible just
watching the LED. Commenting out the message made the LED flicker much
more regularly -- even when shorting the pins.