Improve timing of lighting LED.
The previous code tried to wake regularly (@20 Hz) to sample the sump
pump activity. Before doing anything, it lights the LED. I tried
stretching the time the LED was on by calling `nanosleep` after the
task finished all its work. Unfortunately, if sending a state change
to the network caused a delay, the stretch timer would pass the next
sample time. The loop would wake immediately and process the input
again and then delay again. The result is the LED would appear to stay
on constantly (in fact it was shutting off briefly.)

This commit takes the wake time and adds 15ms and delays again so that
the timing is tied to absolute time instead of a floating, relative
time.

The LED now flickers much more reguarly.