Use `CLOCK_MONOTONIC`.
The LED seemed to flash regularly for weeks and then one day it would
flash normally for a while and then be mostly off for a while. It took
a little time before I realized I enabled `ntpd`. As `ntpd` adjusted
the sysem time, the timing of the LED was being effected.

This commit uses the `CLOCK_MONOTONIC` clock which simply advances at
a constant rate. The data we send to a client no longer as an absolute
timestamp but, instead has relative ones.