:construction: major shift in direction
For this application, I always felt the RaspberryPi 3+ was overkill. And
it was. The recent product, the RaspberryPi Pico W, is a microcontroller
with WiFi connectivity. There's no OS, so you can get real-time
monitoring.

This branch chronicles the development of the software on the
microcontroller. Once this is done, I'll be able to use the RaspberryPi
3 for other, more general things.