:sparkles: blinks the LED
This blinks the onboard LED, just like the demo. However, this version
is a "minimal" codebase. I copied the code from the demo, and then added
`use` statements and added crates until it compiled and ran. The example
pulled in many, many crates because the examples tested out all the
peripherals.

In this version, I also used a `Ticker` so that the timestamps are
rock-steady.

I pulled in the firmware from the Embassy project since they're under a
permissive license.