:zap: switch to "performnce" mode for WiFi
When using MicroPython, I noticed large latencies in network responses
when the interface was in "power save" mode. Turning on "performance"
kept the latencies in the millisecond range.

Since this project will be using network packet timing to timestamp the
states of the sump pump (and I was using 50ms as the base polling speed
in the previous implementation), we can't have WiFi latencies get up
100s of milliseconds (which I saw!)