:sparkles: add a "keepalive" handler
The TCP socket is already configured to to do a transport-level
keep-alive. This works well in letting the microcontroller detect when a
client disconnects unexpectedly. This commit also sets up a timer which
emits a keepalive message when there's no activity on the socket for an
amount of time. This does two things: clients that aren't using TCP
keep alives can use these higher level messages are track when the
controller isn't responding (power outage, bug, etc.) It also lets the
client associate the controller's timestamps with the client's system
time.